Adds the ability to add images to an existing run (#443)
* SQL update functions

* Initial commit to add image mode

- Run configuration checks.
- Constructing the sources_df ready for new association.

* Working version

- Outputs need to be checked!
- Parallel currently broken.

* Attempts to fix parallel association

* Removed unused import

* Fixed typo

* Removed unused import

* Bug fixing id_a > id_b

* Ensured measurement pairs are in date (a < b) order

* Add loading

* Fixed datetime merging

* Implement SQL update

* Plug in sql update

* Fixed running when status is INI

* Tests for adding an image

* Relations fix, parallel fix and ERR status fix

* Fixed monitoring measurement pairs in add image mode

* Fixed path to use pre-defined variable

* Add tests for add image

* Style

* Updated UI with add image mode

- Also fixed copy backup parquets from UI logic.

* Add config files for test

* Fixed bug in tests - doesn't re-run same run again.

* WIP commit

* Fixed basic association

* Added bootstrap4toggle to gulp merge return

* Condensed basic appends to one

* Added back reset_index

* Add image test now uses full dataset.

* Added more tests

* Changed config to only add images at a later date

* Stopped re-running the pipeline before every test

* Fix reconstruction sources_df ra and dec

* Tests changed to 3 images

* Fix reconstruct sources_df position errors

* Fixed advanced relation issue

- The d2d on skyc2 was being added incorrectly making the assignment of the d2d value not correct in relation cases.
- d2d threshold cut is now applied after the merge.

* Minor tidy

* Edited regression tests for advanced

* Update tests, contributing, and changelog

* Broke up lines

* Broke up lines

* Fix typo

* Engrish

* Add forced reshape

* Fix add mode forced extractions

- Detection image was wrongly assigned in add mode in the ideal sources df.
- Fixed by dropping all the forced measurements when calculating the missing_sources_df.

* Added comment

* Change padding from minutes to run id

* Removed unused code

* Fix for catching one image sky region groups

- Effectively this catches running the pipeline with only one image.

* Updated comments

* Docstrings and epoch_based flag fix on complete-rerun

- Added docstrings to new functions.
- Fixed issue where epoch_based flag was not being reset on complete-reruns.

* Adding typing

* Refactored tests, added some for add image

* Init file so test directory is found

* Add remove files if run not in database

* Extra tests

* tests

* Fix parallel add mode

- Source ids were not be corrected properly in parallel add mode - fixed logic.
- Also made sure that images are associated in date order when in parallel mode.

* Tidy code

* Removed unused code in tests

* Edit cloudstor link to cutouts

* open files with pandas instead of dask

* Forced epoch test fix

* Deruiter for normal

* Removed unused complete-rerun flag

* Add deruiter to add image test

* Add deruiter to parallel test.

* Add deruiter to parallel add image tests

* Add deruiter to forced tests

* Add deruiter to epoch tests

* Add deruiter to epoch parallel add image tests

* Unify names tests and add deruiter to epoch forced

* Test config

* forced num

* update changelog

* Made add image UI instructions clearer.

* Change order of if else for readbility

* Fix config files for add image deruiter

* Fix adding with one image

- 1 image in add mode was being incorrectly flagged as nothing to add.

* Removed not needed empty add image check

- It is check earlier.

* Add test 2 adds test

* Add configs for add two images

* Forgot config_add2 definition

* Correct directory

* Add command to only run unit tests to contributing

* Apply suggestions from code review


* Update vast_pipeline/pipeline/utils.py

Missed commit in batch commit.


* Review changes and forced test fix

* Added script to download regression test data

* Remove trailing white spaces

* Add regression tests to CICD

* Updated CONTRIBUTING.md

* Second round of review comments

* Apply suggestions from code review


* Add type annotations and clean up code

* Third round of review comments

* Update vast_pipeline/pipeline/utils.py


* Added missing parameters to docstring

* Merge loops

* Add missing :

* Another missing :

* Fourth round of review comments

* Fix typings

* Add type annotations to tests

* Remove extra space

* Add code to generate config file for testing

* Generate config files on the fly

* Change fully to generating config files on the fly.
Changes to template are back compatible. Only the tests give epoch_mode

* Apply suggestions from code review


* Update vast_pipeline/pipeline/utils.py


* Code review edit

* Remove old backup files on full re-run

* run cicd on python 3.6 and 3.8 as well

* CICD -- turn off fail-fast and specify exact python versions

* bugfix testsuite

* Delete test directories if they exist, before trying to create them

* Save backup of config.py to match bak files

* Update vast_pipeline/tests/test_regression/test_add_image.py

+++ vast_pipeline/pipeline/loading.py
@@ -1,9 +1,11 @@
 import os
 import logging
+import numpy as np
 import pandas as pd
 
-from django.db import transaction
+from typing import List, Optional
 from itertools import islice
+from django.db import transaction, connection, models
 
 from vast_pipeline.image.main import SelavyImage
 from vast_pipeline.pipeline.model_generator import (
@@ -13,7 +15,9 @@ from vast_pipeline.pipeline.model_generator import (
     association_models_generator,
     measurement_pair_models_generator,
 )
-from vast_pipeline.models import Association, Measurement, Source, RelatedSource, MeasurementPair
+from vast_pipeline.models import (
+    Association, Measurement, Source, RelatedSource, MeasurementPair, Run
+)
 from vast_pipeline.pipeline.utils import (
     get_create_img, get_create_img_band
 )
@@ -149,14 +153,31 @@ def make_upload_images(paths, config, pipeline_run):
     return images, skyregs_df
 
 
-def make_upload_sources(sources_df, pipeline_run):
+def make_upload_sources(sources_df: pd.DataFrame, pipeline_run: Run,
+    add_mode: bool = False) -> pd.DataFrame:
     '''
-    delete previous sources for given pipeline run and bulk upload
+    Delete previous sources for given pipeline run and bulk upload
     new found sources as well as related sources
+
+    Parameters
+    ----------
+    sources_df : pd.DataFrame
+        Holds the measurements associated into sources. The output of of the
+        association step.
+    pipeline_run : Run
+        The pipeline Run object.
+    add_mode : bool
+        Whether the pipeline is running in add image mode.
+
+    Returns
+    -------
+    sources_df : pd.DataFrame
+        The input dataframe with the 'id' column added.
     '''
     # create sources in DB
     with transaction.atomic():
-        if Source.objects.filter(run=pipeline_run).exists():
+        if (add_mode is False and
+                Source.objects.filter(run=pipeline_run).exists()):
             logger.info('Removing objects from previous pipeline run')
             n_del, detail_del = (
                 Source.objects.filter(run=pipeline_run).delete()
@@ -210,3 +231,111 @@ def make_upload_measurement_pairs(measurement_pairs_df):
     )
     measurement_pairs_df["id"] = meas_pair_dj_ids
     return measurement_pairs_df
+
+
+def update_sources(
+    sources_df: pd.DataFrame, batch_size: int = 10_000) -> pd.DataFrame:
+    '''
+    Update database using SQL code. This function opens one connection to the
+    database, and closes it after the update is done.
+
+    Parameters
+    ----------
+    sources_df : pd.DataFrame
+        DataFrame containing the new data to be uploaded to the database. The
+        columns to be updated need to have the same headers between the df and
+        the table in the database.
+    batch_size : int
+        The df rows are broken into chunks, each chunk is executed in a
+        separate SQL command, batch_size determines the maximum size of the
+        chunk.
+
+    Returns
+    -------
+    sources_df : pd.DataFrame
+        DataFrame containing the new data to be uploaded to the database.
+    '''
+    # Get all possible columns from the model
+    all_source_table_cols = [
+        fld.attname for fld in Source._meta.get_fields()
+        if getattr(fld, 'attname', None) is not None
+    ]
+
+    # Filter to those present in sources_df
+    columns = [
+        col for col in all_source_table_cols if col in sources_df.columns
+    ]
+
+    sources_df['id'] = sources_df.index.values
+
+    batches = np.ceil(len(sources_df)/batch_size)
+    dfs = np.array_split(sources_df, batches)
+    with connection.cursor() as cursor:
+        for df_batch in dfs:
+            SQL_comm = SQL_update(
+                df_batch, Source, index='id', columns=columns
+            )
+            cursor.execute(SQL_comm)
+
+    return sources_df
+
+
+def SQL_update(
+    df: pd.DataFrame, model: models.Model, index: Optional[str] = None,
+    columns: Optional[List[str]] = None
+) -> str:
+    '''
+    Generate SQL code required to update database.
+
+    Parameters
+    ----------
+    df : pd.DataFrame
+        DataFrame containing the new data to be uploaded to the database. The
+        columns to be updated need to have the same headers between the df and
+        the table in the database.
+    model : Model
+        The model that is being updated.
+    index : str or None
+        Header of the column to join on, determines which rows in the different
+        tables match. If None, then use the primary key column.
+    columns : List[str] or None
+        The column headers of the columns to be updated. If None, updates all
+        columns except the index column.
+
+    Returns
+    -------
+    SQL_comm : str
+        The SQL command to update the database.
+    '''
+    # set index and columns if None
+    if index is None:
+        index = model._meta.pk.name
+    if columns is None:
+        columns = df.columns.tolist()
+        columns.remove(index)
+
+    # get names
+    table = model._meta.db_table
+    new_columns = ', '.join('new_'+c for c in columns)
+    set_columns = ', '.join(c+'=new_'+c for c in columns)
+
+    # get index values and new values
+    column_headers = [index]
+    column_headers.extend(columns)
+    data_arr = df[column_headers].to_numpy()
+    values = []
+    for row in data_arr:
+        val_row = '(' + ', '.join(f'{val}' for val in row) + ')'
+        values.append(val_row)
+    values = ', '.join(values)
+
+    # update database
+    SQL_comm = f"""
+        UPDATE {table}
+        SET {set_columns}
+        FROM (VALUES {values})
+        AS new_values (index_col, {new_columns})
+        WHERE {index}=index_col;
+    """
+
+    return SQL_comm
